changing float to double
changing float to double
and fixing minor bug

+++ main/AreaBounds.cpp
@@ -1,12 +1,12 @@
  #include "AreaBounds.h"
  
-  Area::Area(float xmin_,float xmax_,float ymin_,float ymax_,bool Free_)
+  Area::Area(double xmin_,double xmax_,double ymin_,double ymax_,bool Free_)
   {
     Free=Free_;
     if(Free)
     {
-      xmax=ymax= std::numeric_limits<float>::max();
-      xmin=ymin= std::numeric_limits<float>::min();
+      xmax=ymax= std::numeric_limits<double>::max();
+      xmin=ymin= std::numeric_limits<double>::min();
     }
     else
     {
